Breakup reload, which generally is a long running test
