+++ plugins/wiggle/src/index.test.ts
@@ -3,7 +3,6 @@ import { getSnapshot } from 'mobx-state-tree'
 import ThisPlugin from '.'
 
 test('plugin in a stock JBrowse', () => {
-  console.warn = jest.fn()
   const pluginManager = new PluginManager([new ThisPlugin()])
   pluginManager.createPluggableElements()
   pluginManager.configure()
