+++ products/jbrowse-web/src/tests/BreakpointSplitView.test.tsx
@@ -2,7 +2,7 @@ import '@testing-library/jest-dom/extend-expect'
 import { waitFor } from '@testing-library/react'
 
 import breakpointConfig from '../../test_data/breakpoint/config.json'
-import { createView, doBeforeEach, setup } from './util'
+import { createView, doBeforeEach, mockConsoleWarn, setup } from './util'
 
 setup()
 
@@ -13,15 +13,16 @@ beforeEach(() => {
 const delay = { timeout: 10000 }
 
 test('breakpoint split view', async () => {
-  console.warn = jest.fn()
-  const { findByTestId, queryAllByTestId } = createView(breakpointConfig)
+  await mockConsoleWarn(async () => {
+    const { findByTestId, queryAllByTestId } = createView(breakpointConfig)
 
-  // the breakpoint could be partially loaded so explicitly wait for two items
-  await waitFor(() => expect(queryAllByTestId('r1').length).toBe(2), delay)
+    // the breakpoint could be partially loaded so explicitly wait for two items
+    await waitFor(() => expect(queryAllByTestId('r1').length).toBe(2), delay)
 
-  expect(
-    await findByTestId('pacbio_hg002_breakpoints-loaded'),
-  ).toMatchSnapshot()
+    expect(
+      await findByTestId('pacbio_hg002_breakpoints-loaded'),
+    ).toMatchSnapshot()
 
-  expect(await findByTestId('pacbio_vcf-loaded')).toMatchSnapshot()
+    expect(await findByTestId('pacbio_vcf-loaded')).toMatchSnapshot()
+  })
 }, 10000)
