+++ products/jbrowse-web/src/tests/ErrorConditions.test.tsx
@@ -1,4 +1,4 @@
-import { createView, doBeforeEach } from './util'
+import { createView, doBeforeEach, mockConsole } from './util'
 import chromeSizesConfig from '../../test_data/config_chrom_sizes_test.json'
 import wrongAssemblyTest from '../../test_data/wrong_assembly.json'
 
@@ -9,13 +9,16 @@ beforeEach(() => {
 })
 
 test('404 sequence file', async () => {
-  console.error = jest.fn()
-  const { findAllByText } = createView(chromeSizesConfig)
-  await findAllByText(/HTTP 404 fetching grape.chrom.sizes.nonexist/)
+  await mockConsole(async () => {
+    const { findAllByText } = createView(chromeSizesConfig)
+    await findAllByText(/HTTP 404 fetching grape.chrom.sizes.nonexist/)
+  })
 })
 
 test('wrong assembly', async () => {
-  const { view, findAllByText } = createView(wrongAssemblyTest)
-  view.showTrack('volvox_wrong_assembly')
-  await findAllByText(/does not match/, {}, delay)
+  await mockConsole(async () => {
+    const { view, findAllByText } = createView(wrongAssemblyTest)
+    view.showTrack('volvox_wrong_assembly')
+    await findAllByText(/does not match/, {}, delay)
+  })
 }, 15000)
