+++ products/jbrowse-web/src/tests/__image_snapshots__/reload-bam-test-tsx-reloads-alignments-track-bam-404-1-snap.png
